first bit of lifting the GOT layout out of `generate_llvm_index`.
This bit takes care of passing the parameters properly and adding the required
fields to the required structs. there is also the lifting of the "reading" of GOT layouts to the function calling `generate_module`.
now we need to extract the generated new GOT layouts from the generated modules (?) and merge them together OR, which would be preferable,
pass a mutex to the GotLayout with abstractions for adding a new index in it.

NOTE: How to do that merge afterwards? How to access that mutex properly?
NOTE: If we return a GeneratedModule and then collect those into a vector of GeneratedModule, we need to merge the GotLayouts at this point
NOTE: ...which is why it probably makes more sense to go with the mutex, this way there is no got layout to return in GeneratedModule?